+++ scripts/apply_finetuning.py
@@ -74,10 +74,10 @@ print("Extracting the pre-trained encoder...")
 
 # Define the inputs with variable sequence length
 input_layer = {
-    'input': tf.keras.Input(shape=(build_seq_len, 1), name='input', dtype=tf.float64),
-    'times': tf.keras.Input(shape=(build_seq_len, 1), name='times', dtype=tf.float64),
-    'band_info': tf.keras.Input(shape=(build_seq_len, 1), name='band_info', dtype=tf.float64),
-    'mask': tf.keras.Input(shape=(build_seq_len,), name='mask', dtype=tf.float64)
+    'input': tf.keras.Input(shape=(build_seq_len, 1), name='input', dtype=tf.float32),
+    'times': tf.keras.Input(shape=(build_seq_len, 1), name='times', dtype=tf.float32),
+    'band_info': tf.keras.Input(shape=(build_seq_len, 1), name='band_info', dtype=tf.float32),
+    'mask': tf.keras.Input(shape=(build_seq_len,), name='mask', dtype=tf.float32)
 }
 
 # --- Break down the forward pass step-by-step ---
